Improved embedded server and Gradle/IDE integration
Added HTTPS support to the embedded Jetty and enabled its configuration through a set of properties.
Improved the integration of the Gradle multi project build with the IDE by introducing a corresponding Gradle task (runAppInIDE) which provides the same runtime environment as for the distribution (correct resolution of config.properties, hazelcast.xml and logback.xml).
Improved .gitignore file to ignore additional generated files and folders.