Added Persistence Support for all Model Objects.
Created an interface and factory class for data manager implementations as well as two initial data managers. One SimpleDataManager with the already existing code and an HazelcastDataManager that will be used to support multi node deployments of the middleware using Hazelcast in the future.
Introduced a new interface IPersistenceProvider that specifies all required methods for the persistence of binary data as well as all managed objects of the middleware.
Therefore, we introduced a new PersistableObject interface that is implemented by all model objects and allows us to access the identifier of the model objects in order to store and load them in corresponding data sources. As a bridge between the persistence functionality provided by IPersistenceProvider implementations and the data management functionality provided by IDataManager implementations we introduced a PersistableHashMap (which extends ConcurrentHashMap). This map allows us to collect and persist all types of model objects while hiding the concrete implementation details.